requirements.txt
requests
numpy
pandas
sqlalchemy
python-dotenv
psycopg2-binary